+++ .idea/appInsightsSettings.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="AppInsightsSettings">
+    <option name="selectedTabId" value="Android Vitals" />
     <option name="tabSettings">
       <map>
         <entry key="Firebase Crashlytics">
@@ -8,7 +9,7 @@
             <InsightsFilterSettings>
               <option name="connection">
                 <ConnectionSetting>
-                  <option name="appId" value="com.example.loginscreen" />
+                  <option name="appId" value="com.example.veteranreccanteen" />
                   <option name="mobileSdkAppId" value="1:299375459032:android:33a625d3d71e07176eb5ad" />
                   <option name="projectId" value="tugas-pbp-uts" />
                   <option name="projectNumber" value="299375459032" />
