+++ app/src/androidTest/java/com/example/veteranreccanteen/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.loginscreen;
+package com.example.veteranreccanteen;
 
 import android.content.Context;
