+++ app/src/google-services.json
@@ -29,7 +29,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:299375459032:android:33a625d3d71e07176eb5ad",
         "android_client_info": {
-          "package_name": "com.example.loginscreen"
+          "package_name": "com.example.veteranreccanteen"
         }
       },
       "oauth_client": [],
