+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.LoginScreen"
+        android:theme="@style/Theme.VCR"
         tools:targetApi="31">
         <activity
             android:name=".SplashScreen"
@@ -28,7 +28,7 @@
             android:name=".Login"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".Homepage"
             android:exported="false">
         </activity>
     </application>
