+++ app/src/main/res/drawable/ic_profile.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="48dp" android:viewportHeight="32" android:viewportWidth="32" android:width="48dp">
+      
+    <path android:fillColor="#9F9F9F" android:fillType="evenOdd" android:pathData="M22.5,23.599C24.642,21.765 26,19.041 26,16C26,10.477 21.523,6 16,6C10.477,6 6,10.477 6,16C6,19.041 7.357,21.765 9.5,23.599C9.828,21.156 11.922,19.273 14.455,19.273H17.545C20.079,19.273 22.171,21.156 22.5,23.599ZM18.87,27.655C24.111,26.368 28,21.638 28,16C28,9.373 22.627,4 16,4C9.373,4 4,9.373 4,16C4,20.213 6.171,23.918 9.455,26.059V26.059C10.691,26.866 12.086,27.45 13.582,27.756C14.363,27.916 15.172,28 16,28M18.87,27.655C17.95,27.88 16.989,28 16,28L18.87,27.655ZM20.545,24.91V24.273C20.545,22.616 19.202,21.273 17.545,21.273H14.455C12.798,21.273 11.455,22.616 11.455,24.273V24.91C12.746,25.57 14.201,25.958 15.742,25.997C15.828,25.999 15.914,26 16,26C16.108,26 16.216,25.998 16.323,25.995C17.84,25.947 19.272,25.561 20.545,24.91ZM16,15.091C17.305,15.091 18.364,14.033 18.364,12.727C18.364,11.422 17.305,10.364 16,10.364C14.695,10.364 13.636,11.422 13.636,12.727C13.636,14.033 14.695,15.091 16,15.091ZM16,17.091C18.41,17.091 20.364,15.137 20.364,12.727C20.364,10.317 18.41,8.364 16,8.364C13.59,8.364 11.636,10.317 11.636,12.727C11.636,15.137 13.59,17.091 16,17.091Z"/>
+    
+</vector>
