+++ app/src/main/res/layout/activity_sign_up.xml
@@ -45,6 +45,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
         android:orientation="vertical"
         android:layout_gravity="center_horizontal">
