+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name" translatable="false">Login Screen</string>
+    <string name="app_name" translatable="false">Veteran Recommendation Canteen</string>
 </resources>
