+++ app/src/test/java/com/example/veteranreccanteen/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.loginscreen;
+package com.example.veteranreccanteen;
 
 import org.junit.Test;
